// frontend/e-learn/backoffice/src/app/models/cours.model.ts
export interface Cours {
  id?: number;
  title?: string;
  description?: string;
  cover?: string;
  video?: string;
  category?: string;
  language?: string;
  level?: string;
  price?: number;
  duration?: string;
  instructor?: string;
  created_at?: string;
  updated_at?: string;
  status?: string;
}